Создаёт store и выносит в него всю логику по созданию стора и загрузке данных(кроме комментариев)
