+++ src/index.js
@@ -1,35 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./components/app/app.jsx";
-import {applyMiddleware, createStore} from "redux";
-import reducer from "./reducer/reducer";
 import {Provider} from "react-redux";
-import {createAPI} from "./api";
-import {composeWithDevTools} from "redux-devtools-extension";
-import thunk from "redux-thunk";
-import {ActionCreator, AuthorizationStatus} from "./reducer/user/user";
+import store from "./reducer/store";
 
-const onUnauthorized = () => {
-  store.dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.NO_AUTH));
+const init = () => {
+  ReactDOM.render(
+      <Provider store={store}>
+        <App/>
+      </Provider>,
+      document.querySelector(`#root`)
+  );
 };
 
-const onAuthError = () => {
-  store.dispatch(ActionCreator.setLoginErrorStatus(true));
-};
-
-const api = createAPI(onUnauthorized, onAuthError);
-
-const store = createStore(
-    reducer,
-    composeWithDevTools(
-        applyMiddleware(thunk.withExtraArgument(api))
-    )
-);
-
-ReactDOM.render(
-    <Provider store={store}>
-      <App/>
-    </Provider>,
-    document.querySelector(`#root`)
-);
-
+init();
